Remove duplication into started at trait.
