+++ app/Filters/RefereeFilters.php
@@ -2,11 +2,11 @@
 
 namespace App\Filters;
 
-use Carbon\Carbon;
 use Illuminate\Support\Str;
 
 class RefereeFilters extends Filters
 {
+    use Concerns\FiltersByStartDate,
     /**
      * Registered filters to operate upon.
      *
@@ -29,27 +29,4 @@ class RefereeFilters extends Filters
         return $this->builder;
     }
 
-    /**
-     * Filter a query to include referees of a specific started at date.
-     *
-     * @param  array  $startedAt
-     * @return \Illuminate\Database\Eloquent\Builder
-     */
-    public function started_at($startedAt)
-    {
-        if (isset($startedAt[1])) {
-            $this->builder->whereHas('employment', function ($query) use ($startedAt) {
-                $query->whereBetween('started_at', [
-                    Carbon::parse($startedAt[0])->toDateString(),
-                    Carbon::parse($startedAt[1])->toDateString()
-                ]);
-            });
-        } elseif (isset($startedAt[0])) {
-            $this->builder->whereHas('employment', function ($query) use ($startedAt) {
-                $query->whereDate('started_at', Carbon::parse($startedAt[0])->toDateString());
-            });
-        }
-
-        return $this->builder;
-    }
 }
